change the data type of mp4/webp size in test cases

+++ examples/qtmegachatapi/chatWindow.cpp
@@ -1318,8 +1318,8 @@ void ChatWindow::onAttachGiphy()
     //giphy data
     const char* srcMp4 = "giphy://media/Wm9XlKG2xIMiVcH4CP/200.mp4?cid=a2a900dl&rid=200.mp4&dom=bWVkaWEyLmdpcGh5LmNvbQ%3D%3D";
     const char* srcWebp = "giphy://media/Wm9XlKG2xIMiVcH4CP/200.webp?cid=a2a900dl&rid=200.webp&dom=bWVkaWEyLmdpcGh5LmNvbQ%3D%3D";
-    long sizeMp4 = 59970;
-    long sizeWebp = 159970;
+    long long sizeMp4 = 59970;
+    long long sizeWebp = 159970;
     int giphyWidth = 200;
     int giphyHeight = 200;
     const char* giphyTitle = "TEST_SendGiphy";
